+++ src/examples/lib_svm_example.cc
@@ -21,7 +21,7 @@ int main(int argc, char* argv[])
 
   size_t numberOfData=trainData.rows();
   size_t numberOfFeatures=trainData.cols();
-  double gamma=1.0/numberOfFeatures;
+  double gamma=1.0/(double)numberOfFeatures;
 
   Kernel kernel{gamma};
   LearningModel learningModel{kernel, numberOfFeatures};
@@ -38,9 +38,9 @@ int main(int argc, char* argv[])
   t=clock()-t;
   printf ("It took me %ld clicks (%f seconds) for predicting.\n",t,((float)t)/CLOCKS_PER_SEC);
 
-  double accuracy=1.0-LossFunction(predictedLabels, trainPair.second)/numberOfData;
+  double accuracy=1.0-LossFunction(predictedLabels, trainPair.second)/(double)numberOfData;
   printf("accuracy = %f%%, (%u / %lu)\n", accuracy*100,
-         (unsigned)(accuracy*numberOfData), numberOfData);
+         (unsigned)(accuracy*(double)numberOfData), numberOfData);
 
   return 0;
 }
